// src/types/app.ts
export interface FileUploadConfig {
  file_size_limit: number;
  batch_count_limit: number;
  file_upload_limit: number;
  image_file_size_limit: number;
  video_file_size_limit: number;
  audio_file_size_limit: number;
  workflow_file_upload_limit: number;
  image_file_batch_limit: number;
  single_chunk_attachment_limit: number;
  attachment_image_file_size_limit: number;
}

export interface FileUploadSettings {
  enabled: boolean;
  allowed_file_types: string[];
  allowed_file_extensions: string[];
  allowed_file_upload_methods: TransferMethod[];
  number_limits: number;
  fileUploadConfig: FileUploadConfig;
}

export interface SpeechToTextSettings {
  enabled: boolean;
}

export interface TextToSpeechSettings {
  enabled: boolean;
  language: string;
  voice: string;
  autoPlay: string;
}

export interface AppParams {
  opening_statement: string;
  suggested_questions: string[];
  suggested_questions_after_answer: { enabled: boolean };
  speech_to_text: SpeechToTextSettings;
  text_to_speech: TextToSpeechSettings;
  retriever_resource: { enabled: boolean };
  annotation_reply: { enabled: boolean };
  more_like_this: { enabled: boolean };
  user_input_form: UserInputFormItem[];
  sensitive_word_avoidance: { enabled: boolean };
  file_upload: FileUploadSettings;
  system_parameters: Record<string, unknown>;
}

export interface PromptConfig {
  prompt_template: string;
  prompt_variables: PromptVariable[];
}

export interface TextTypeFormItem {
  label: string;
  variable: string;
  required: boolean;
  max_length: number;
}

export interface SelectTypeFormItem {
  label: string;
  variable: string;
  required: boolean;
  options: string[];
}
/**
 * User Input Form Item
 */
export type UserInputFormItem =
  | {
      'text-input': TextTypeFormItem;
    }
  | {
      select: SelectTypeFormItem;
    }
  | {
      paragraph: TextTypeFormItem;
    };

export const MessageRatings = ['like', 'dislike', null] as const;
export type MessageRating = (typeof MessageRatings)[number];

export interface Feedbacktype {
  rating: MessageRating;
  content?: string | null;
}

export interface MessageMore {
  time: string;
  tokens: number;
  latency: number | string;
}

export interface IChatItem {
  id: string;
  content: string;
  /**
   * Specific message type
   */
  isAnswer: boolean;
  /**
   * The user feedback result of this message
   */
  feedback?: Feedbacktype;
  /**
   * The admin feedback result of this message
   */
  adminFeedback?: Feedbacktype;
  /**
   * Whether to hide the feedback area
   */
  feedbackDisabled?: boolean;
  /**
   * More information about this message
   */
  more?: MessageMore;
  annotation?: Annotation;
  useCurrentUserAvatar?: boolean;
  isOpeningStatement?: boolean;
  suggestedQuestions?: string[];
  log?: { role: string; text: string }[];
  agent_thoughts?: ThoughtItem[];
  message_files?: VisionFile[];
}

export type ChatItem = IChatItem & {
  isError?: boolean;
  workflow_run_id?: string;
  workflowProcess?: WorkflowProcess;
};

export interface ConversationItem {
  id: string;
  name: string;
  inputs: Record<string, unknown> | null;
  introduction: string;
  suggested_questions?: string[];
}

export interface AppInfo {
  title: string;
  description: string;
  default_language: string;
  copyright?: string;
  privacy_policy?: string;
  disable_session_same_site?: boolean;
}

export enum Resolution {
  low = 'low',
  high = 'high',
}

export enum TransferMethod {
  all = 'all',
  local_file = 'local_file',
  remote_url = 'remote_url',
}

export interface VisionSettings {
  enabled: boolean;
  number_limits: number;
  detail: Resolution;
  transfer_methods: TransferMethod[];
  image_file_size_limit?: number | string;
}

export interface ImageFile {
  type: TransferMethod;
  _id: string;
  fileId: string;
  file?: File;
  progress: number;
  url: string;
  base64Url?: string;
  deleted?: boolean;
}

export interface VisionFile {
  id?: string;
  type: string;
  transfer_method: TransferMethod;
  url: string;
  upload_file_id: string;
  belongs_to?: string;
}

export enum BlockEnum {
  Start = 'start',
  End = 'end',
  Answer = 'answer',
  LLM = 'llm',
  KnowledgeRetrieval = 'knowledge-retrieval',
  QuestionClassifier = 'question-classifier',
  IfElse = 'if-else',
  Code = 'code',
  TemplateTransform = 'template-transform',
  HttpRequest = 'http-request',
  VariableAssigner = 'variable-assigner',
  Tool = 'tool',
}

export interface NodeTracing {
  id: string;
  index: number;
  predecessor_node_id: string;
  node_id: string;
  node_type: BlockEnum;
  title: string;
  inputs: unknown;
  process_data: unknown;
  outputs?: unknown;
  status: string;
  error?: string;
  elapsed_time: number;
  execution_metadata: {
    total_tokens: number;
    total_price: number;
    currency: string;
  };
  created_at: number;
  created_by: {
    id: string;
    name: string;
    email: string;
  };
  finished_at: number;
  extras?: unknown;
  expand?: boolean; // for UI
}

export enum NodeRunningStatus {
  NotStart = 'not-start',
  Waiting = 'waiting',
  Running = 'running',
  Succeeded = 'succeeded',
  Failed = 'failed',
}

export enum WorkflowRunningStatus {
  Waiting = 'waiting',
  Running = 'running',
  Succeeded = 'succeeded',
  Failed = 'failed',
  Stopped = 'stopped',
}

export interface WorkflowProcess {
  status: WorkflowRunningStatus;
  tracing: NodeTracing[];
  expand?: boolean; // for UI
}